requirements.txt
Django~=5.1.1
python-docx~=1.1.2
nepali~=1.1.1